edits to include APY

+++ public_html/Project/sql/010_create_table_system_properties.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS  `System_Properties`
 (
     `id` int auto_increment not null,
     `name` varchar(7) not null,
-    `value` int default 0,
+    `value` DECIMAL(3,2) default 0.0,
     `created` timestamp default current_timestamp,
     `modified` timestamp default current_timestamp on update current_timestamp,
     PRIMARY KEY (`id`)
